requirements.txt
gradio
smolagents
yaml